Split out Version and ModuleSpec tests into spec files of their own
